modified the file struture for some refactoring ideas
