+++ triton-client/package.json
@@ -12,10 +12,10 @@
         "url": "https://github.com/c3g/triton/issues"
     },
     "scripts": {
-        "start": "NODE_ENV=development react-scripts start",
-        "build": "NODE_ENV=production react-scripts build",
-        "test": "react-scripts test",
-        "eject": "react-scripts eject"
+        "start": "NODE_ENV=development craco start",
+        "build": "NODE_ENV=production craco build",
+        "test": "craco test",
+        "eject": "craco eject"
     },
     "browserslist": {
         "production": [
@@ -38,6 +38,11 @@
         "redux-logger": "^3.0.6"
     },
     "devDependencies": {
+        "@babel/core": "^7.24.7",
+        "@babel/preset-env": "^7.24.7",
+        "@babel/preset-react": "^7.24.7",
+        "@babel/preset-typescript": "^7.24.7",
+        "@craco/craco": "^7.1.0",
         "@reduxjs/toolkit": "^1.8.3",
         "@testing-library/jest-dom": "^5.16.4",
         "@testing-library/react": "^13.3.0",
@@ -47,10 +52,14 @@
         "@types/react": "^18.0.15",
         "@types/react-dom": "^18.0.6",
         "@types/redux-logger": "^3.0.9",
+        "babel-loader": "^9.1.3",
+        "html-webpack-plugin": "^5.6.0",
         "node-sass": "^7.0.3",
         "react-scripts": "5.0.1",
         "sass": "^1.54.3",
         "typescript": "^4.7.4",
-        "webpack": "^5.88.2"
+        "webpack": "^5.92.1",
+        "webpack-cli": "^5.1.4",
+        "webpack-dev-server": "^5.0.4"
     }
 }
