+++ triton-client/src/components/MGCHeader/index.tsx
@@ -1,0 +1,10 @@
+import genomeLogo from "@static/genome-logo.jpg"
+
+export const MGCHeader = () => {
+    return <div className="mgc-header">
+        <img
+            alt="McGill Genome Center"
+            height="40px"
+            src={genomeLogo} />
+    </div>
+}
