+++ triton-client/src/components/ProjectCardList/index.scss
@@ -1,0 +1,1 @@
+// the components styling
