+++ triton-client/src/components/ProjectCardList/interfaces.ts
@@ -1,0 +1,7 @@
+import { TritonProject } from "../../api/api-types"
+
+export interface IProjectCardListProps {
+    projects: IProjectCard[]
+}
+
+export interface IProjectCard extends TritonProject {}
