+++ triton-client/src/components/index.ts
@@ -1,0 +1,7 @@
+import ProjectCardList from "./ProjectCardList"
+import { MGCHeader } from "./MGCHeader"
+
+export {
+    ProjectCardList,
+    MGCHeader
+}
